Added support for option groups when parsing plist files.
This is quite simplified for the moment being; the code just checks whether a value for certain key is a dictionary and handles the keys in there as any other keys. This means it doesn't matter whether the name of the dictionary key actually corresponds to one of the names registered to `GBOptions` and/or `GBCommandLineParser`.